+++ gusto/src/main/java/com/umc/gusto/global/auth/model/OAuth2UserWithAuthority.java
@@ -1,4 +1,4 @@
-package com.umc.gusto.domain.user.model;
+package com.umc.gusto.global.auth.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
